create HPP file
